requirements.txt
flask
requests
python-dotenv
xai-sdk
waitress
gunicorn